+++ packages/api/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @atproto/api
 
+## 0.12.8
+
+### Patch Changes
+
+- [`58f719cc1`](https://github.com/bluesky-social/atproto/commit/58f719cc1c8d0ebd5ad7cf11221372b671cd7857) Thanks [@devinivy](https://github.com/devinivy)! - Add grandparent author to feed item reply ref
+
 ## 0.12.7
 
 ### Patch Changes